Tweak Cylon Leader motive counts
Change the formula used on the Image Data pane so we only show a count
of 2 for Cylon Leader motives before the sleeper phase, since the
current behaviour of always showing a count of 4 has led to a number of
questions.

This replaces the current
=Loyalty.<Cell>
formula with the more complex:
=IF(AND('Game State'.<Cell>="L",'Destination Cards'.D6<4),2,Loyalty.<Cell>)

Since Pegasus agendas use 'LS' and 'LH' for leader loyalties, this
check handles the difference between Pegasus Agendas and Daybreak
Motives with no further checks.